Remove circular import
